+++ tests/test_exceptions.py
@@ -1,4 +1,5 @@
 import pytest
+from tests.conftest import REQUEST_EXCEPTION_MESSAGE
 
 from py42.exceptions import Py42BadRequestError
 from py42.exceptions import Py42ForbiddenError
@@ -19,7 +20,7 @@ class TestPy42Errors(object):
 
     def test_raise_py42_error_raises_unauthorized_error(self, error_response):
         error_response.response.status_code = 401
-        with pytest.raises(Py42UnauthorizedError):
+        with pytest.raises(Py42UnauthorizedError, match=REQUEST_EXCEPTION_MESSAGE):
             raise_py42_error(error_response)
 
     def test_raise_py42_error_raises_MFA_required_error(self, error_response):
@@ -27,6 +28,10 @@ class TestPy42Errors(object):
         error_response.response.text = (
             '{"error":[{"primaryErrorKey":"TIME_BASED_ONE_TIME_PASSWORD_REQUIRED"}]}'
         )
+        with pytest.raises(Py42MFARequiredError):
+            raise_py42_error(error_response)
+
+        error_response.response.text = '{"error":[{"primaryErrorKey":"TOTP_AUTH_CONFIGURATION_REQUIRED_FOR_USER"}]}'
         with pytest.raises(Py42MFARequiredError):
             raise_py42_error(error_response)